+++ package.json
@@ -26,6 +26,7 @@
     "@types/uuid": "^9.0.1",
     "autoprefixer": "^10.4.13",
     "axios": "^1.3.3",
+    "file-saver": "^2.0.5",
     "mui-image": "^1.0.7",
     "next": "^13.2.3",
     "react": "^18.2.0",
@@ -43,6 +44,7 @@
     "yup": "^1.0.0"
   },
   "devDependencies": {
+    "@types/file-saver": "^2.0.5",
     "@types/mui-image": "^1.0.1",
     "@types/node": "^18.14.6",
     "@types/react": "^18.0.28",
